[bugfix] topo panel, links is incomplete when using auto group by tag in datasource
**Phenomenon and reproduction steps**

none

**Root cause and solution**

- type mismatch in condition judgment

**Impactions**

none

**Test method**

none

**Affected branch(es)**

- main

**Checklist**

- [ ] Dependencies update required
- [ ] Common bug (similar problem in other repo)
